refactor: reorganize files and reuse lander module
