+++ index.html
@@ -85,7 +85,8 @@
       <button id="closeCredits" aria-label="Close credits">Chiudi</button>
     </div>
   </div>
-  <!-- Game logic and drawing script -->
-  <script src="script.js"></script>
+  <!-- Core lander physics and game controller -->
+  <script src="src/lander.js"></script>
+  <script src="src/game.js"></script>
 </body>
 </html>
